+++ plugins/destination/plugin_testing_overwrite.go
@@ -1,0 +1,93 @@
+package destination
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/cloudquery/plugin-sdk/schema"
+	"github.com/cloudquery/plugin-sdk/specs"
+	"github.com/cloudquery/plugin-sdk/testdata"
+	"github.com/google/uuid"
+	"github.com/rs/zerolog"
+)
+
+func (*PluginTestSuite) destinationPluginTestWriteOverwrite(ctx context.Context, p *Plugin, logger zerolog.Logger, spec specs.Destination) error {
+	spec.WriteMode = specs.WriteModeOverwrite
+	if err := p.Init(ctx, logger, spec); err != nil {
+		return fmt.Errorf("failed to init plugin: %w", err)
+	}
+	tableName := fmt.Sprintf("cq_test_write_overwrite_%d", time.Now().Unix())
+	table := testdata.TestTable(tableName)
+	syncTime := time.Now().UTC().Round(1 * time.Second)
+	tables := []*schema.Table{
+		table,
+	}
+	if err := p.Migrate(ctx, tables); err != nil {
+		return fmt.Errorf("failed to migrate tables: %w", err)
+	}
+
+	sourceName := "testOverwriteSource" + uuid.NewString()
+	sourceSpec := specs.Source{
+		Name: sourceName,
+	}
+
+	resources := createTestResources(table, sourceName, syncTime, 2)
+	if err := p.writeAll(ctx, sourceSpec, tables, syncTime, resources); err != nil {
+		return fmt.Errorf("failed to write all: %w", err)
+	}
+	sortResources(table, resources)
+
+	resourcesRead, err := p.readAll(ctx, table, sourceName)
+	if err != nil {
+		return fmt.Errorf("failed to read all: %w", err)
+	}
+	sortCQTypes(table, resourcesRead)
+
+	if len(resourcesRead) != 2 {
+		return fmt.Errorf("expected 2 resources, got %d", len(resourcesRead))
+	}
+
+	if diff := resources[0].Data.Diff(resourcesRead[0]); diff != "" {
+		return fmt.Errorf("expected first resource diff: %s", diff)
+	}
+
+	if diff := resources[1].Data.Diff(resourcesRead[1]); diff != "" {
+		return fmt.Errorf("expected second resource diff: %s", diff)
+	}
+
+	secondSyncTime := syncTime.Add(time.Second).UTC()
+
+	// copy first resource but update the sync time
+	updatedResource := schema.DestinationResource{
+		TableName: table.Name,
+		Data:      make(schema.CQTypes, len(resources[0].Data)),
+	}
+	copy(updatedResource.Data, resources[0].Data)
+	_ = updatedResource.Data[1].Set(secondSyncTime)
+
+	// write second time
+	if err := p.writeOne(ctx, sourceSpec, tables, secondSyncTime, updatedResource); err != nil {
+		return fmt.Errorf("failed to write one second time: %w", err)
+	}
+
+	resourcesRead, err = p.readAll(ctx, table, sourceName)
+	if err != nil {
+		return fmt.Errorf("failed to read all second time: %w", err)
+	}
+	sortCQTypes(table, resourcesRead)
+
+	if len(resourcesRead) != 2 {
+		return fmt.Errorf("after overwrite expected 2 resources, got %d", len(resourcesRead))
+	}
+
+	if diff := resources[1].Data.Diff(resourcesRead[0]); diff != "" {
+		return fmt.Errorf("after overwrite expected first resource diff: %s", diff)
+	}
+
+	if diff := updatedResource.Data.Diff(resourcesRead[1]); diff != "" {
+		return fmt.Errorf("after overwrite expected second resource diff: %s", diff)
+	}
+
+	return nil
+}
